Created server.js
Server.js serves to create a server to store database info such as reports and user data, into a remote MongoDB database. To access this database, use this link: https://www.mongodb.com/atlas
And create an account. Let me know when you create an account so that I can add you to the database and project.