Update README.md
I added a couple of features to this backend API. I added an API tester to test report submission. There's also a real-time status checker that updates every second, and includes an uptime timer. Finally, the README is updated to reflect these changes, as well as including instructions for those that are interested in deploying through Vercel.